Rename ForwardDefines.h to FrameworkFwd.h

+++ OgreRenderingModule/Renderer.cpp
@@ -1546,7 +1546,7 @@ namespace OgreRenderer
 
         if(soft_shadow)
         {
-            for(int i=0;i<shadowTextureCount;i++)
+            for(size_t i=0;i<shadowTextureCount;i++)
             {
                 GaussianListener* gaussianListener = new GaussianListener(); 
                 Ogre::TexturePtr shadowTex = sceneManager->getShadowTexture(0);
